Correction of remarks
