--- src/components/ashesOfWar/AshesOfWar.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const AshesOfWar = () => {
-  return (
-    <div>
-      <h1>Ashes of War</h1>
-    </div>
-  );
-};
