--- src/components/bosses/Bosses.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const Bosses = () => {
-  return (
-    <div>
-      <h1>Bosses</h1>
-    </div>
-  );
-};
