--- src/components/bosses/index.ts
@@ -1,1 +1,0 @@
-export { Bosses } from "./Bosses";
