--- src/components/classes/Classes.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const Classes = () => {
-  return (
-    <div>
-      <h1>classes</h1>
-    </div>
-  );
-};
