--- src/components/classes/index.ts
@@ -1,1 +1,0 @@
-export { Classes } from "./Classes";
