--- src/components/creatures/Creatures.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const Creatures = () => {
-  return (
-    <div>
-      <h1>Creatures</h1>
-    </div>
-  );
-};
