--- src/components/creatures/index.ts
@@ -1,1 +1,0 @@
-export { Creatures } from "./Creatures";
