--- src/components/incantations/Incantations.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const Incantations = () => {
-  return (
-    <div>
-      <h1>Incatations</h1>
-    </div>
-  );
-};
