--- src/components/location/Location.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const Location = () => {
-  return (
-    <div>
-      <h1>Location</h1>
-    </div>
-  );
-};
