--- src/components/location/index.ts
@@ -1,1 +1,0 @@
-export { Location } from './Location';
