--- src/components/npcs/NPCs.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-export const NPCs = () => {
-  return (
-    <div>
-      <h1>NPCs</h1>
-    </div>
-  );
-};
-
-
