+++ src/components/searchName/index.ts
@@ -1,0 +1,1 @@
+export { SearchName } from './SearchName';
