--- src/components/shields/Shields.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const Shields = () => {
-  return (
-    <div>
-      <h1>Shields</h1>
-    </div>
-  );
-};
