--- src/components/sorceries/Sorceries.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const Sorceries = () => {
-  return (
-    <div>
-      <h1>Sorceries </h1>
-    </div>
-  );
-};
