--- src/components/sorceries/index.ts
@@ -1,1 +1,0 @@
-export { Sorceries } from './Sorceries';
