--- src/components/spirits/Spirits.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const Spirits = () => {
-  return (
-    <div>
-      <h1>Spirits</h1>
-    </div>
-  );
-};
