--- src/components/talismans/Talismans.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const Talismans = () => {
-  return (
-    <div>
-      <h1>Talismans</h1>
-    </div>
-  );
-};
