--- src/components/weapons/Weapons.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const Weapons = () => {
-  return (
-    <div>
-      <h1>Weapons</h1>
-    </div>
-  );
-};
