--- src/components/weapons/index.ts
@@ -1,1 +1,0 @@
-export { Weapons } from './Weapons';
