+++ src/index.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import { setupStore } from 'BLL/redux';
 import App from './App';
 import './index.css';
+import { setupStore } from 'reduxLayer/store';
 
 const store = setupStore();
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
