--- src/logic/index.ts
@@ -1,13 +1,0 @@
-export { ItemsContext } from './contect/ItemsContext';
-export { DebounceSearch } from './debounceSearch/DebounceSearch';
-export { options } from './options/options';
-
-export {
-  countItems,
-  enteredSearch,
-  filteredMainTitles,
-} from './universalFunction/universals';
-
-export { useCustomParams } from './customHooks/customHooks';
-export { titlesForSearch } from './universalFunction/titlesForSearch';
-export type { TitleSearch } from './universalFunction/titlesForSearch';
