--- src/logic/options/options.ts
@@ -1,62 +1,0 @@
-export const options = [
-  {
-    value: 'ammos',
-    label: 'ammos',
-  },
-  {
-    value: 'armors',
-    label: 'armors',
-  },
-  {
-    value: 'ashes',
-    label: 'ashes',
-  },
-  {
-    value: 'bosses',
-    label: 'bosses',
-  },
-  {
-    value: 'creatures',
-    label: 'creatures',
-  },
-  {
-    value: 'items',
-    label: 'items',
-  },
-  {
-    value: 'classes',
-    label: 'classes',
-  },
-  {
-    value: 'incantations',
-    label: 'incantations',
-  },
-  {
-    value: 'locations',
-    label: 'locations',
-  },
-  {
-    value: 'NPCs',
-    label: 'NPCs',
-  },
-  {
-    value: 'shields',
-    label: 'shields',
-  },
-  {
-    value: 'sorceries',
-    label: 'sorceries',
-  },
-  {
-    value: 'spirits',
-    label: 'spirits',
-  },
-  {
-    value: 'talismans',
-    label: 'talismans',
-  },
-  {
-    value: 'weapons',
-    label: 'weapons',
-  },
-];
