--- src/logic/universalFunction/titlesForSearch.ts
@@ -1,21 +1,0 @@
-export type TitleSearch = {
-  title: string;
-};
-
-export const titlesForSearch: TitleSearch[] = [
-  { title: 'ammos' },
-  { title: 'items' },
-  { title: 'armors' },
-  { title: 'ashes' },
-  { title: 'bosses' },
-  { title: 'classes' },
-  { title: 'creatures' },
-  { title: 'incantations' },
-  { title: 'locations' },
-  { title: 'npcs' },
-  { title: 'shields' },
-  { title: 'sorceries' },
-  { title: 'spirits' },
-  { title: 'talismans' },
-  { title: 'weapons' },
-];
