+++ src/types/index.ts
@@ -1,0 +1,1 @@
+export type { ResponseItems, ResponseItem, ShortInfoForItem, Items } from './types';
