+++ src/utilities/index.ts
@@ -1,0 +1,12 @@
+export { ItemsContext } from 'utilities/contect/ItemsContext';
+export { options } from 'utilities/options/options';
+
+export {
+  countItems,
+  enteredSearch,
+  filteredMainTitles,
+} from 'utilities/universalFunction/universals';
+
+export { useCustomParams } from 'utilities/customHooks/customHooks';
+export { titlesForSearch } from 'utilities/universalFunction/titlesForSearch';
+export type { TitleSearch } from 'utilities/universalFunction/titlesForSearch';
