+++ src/utilities/options/options.ts
@@ -1,0 +1,18 @@
+export const options = [
+  {
+    value: 'ammos',
+    label: 'ammos',
+  },
+  {
+    value: 'armors',
+    label: 'armors',
+  },
+  {
+    value: 'ashes',
+    label: 'ashes',
+  },
+  {
+    value: 'items',
+    label: 'items',
+  },
+];
