+++ src/utilities/universalFunction/titlesForSearch.ts
@@ -1,0 +1,9 @@
+export type TitleSearch = {
+  title: string;
+};
+
+export const titlesForSearch: TitleSearch[] = [
+  { title: 'ammos' },
+  { title: 'items' },
+  { title: 'armors' },
+];
